+++ packages/utils/src/graphql/index.ts
@@ -2,3 +2,4 @@ export * from "./graphql";
 export * from "./directive";
 export * from "./formatter";
 export * from "./group";
+export * from "./relation";
